Separate test config from test suits
