+++ src/__tests__/Label.test.tsx
@@ -1,0 +1,88 @@
+// @ts-nocheck
+import React from 'react';
+import { render, dataMock, expectedNormalizedDataMock } from './testUtils';
+
+describe('Label', () => {
+  describe('"label"', () => {
+    describe('true', () => {
+      it('renders 3 <text> elements with expected text and "fill" attribute', () => {
+        const { container } = render({ label: true });
+        const labels = container.querySelectorAll('text');
+
+        expect(labels.length).toBe(dataMock.length);
+        labels.forEach((label, index) => {
+          expect(label).toHaveTextContent(`${dataMock[index].value}`);
+          expect(label).toHaveAttribute('fill', dataMock[index].color);
+        });
+      });
+    });
+
+    describe('provided as function returning a value', () => {
+      it('renders 3 <text> elements with custom content', () => {
+        const { container } = render({
+          label: (props) => props.dataIndex,
+        });
+
+        container.querySelectorAll('text').forEach((label, index) => {
+          expect(label).toHaveTextContent(index);
+        });
+      });
+
+      it('provided function receive expected "props" object', () => {
+        const labelMock = jest.fn();
+        render({
+          label: labelMock,
+        });
+
+        const expected = {
+          key: expect.any(String),
+          ...expectedNormalizedDataMock,
+        };
+        expect(labelMock).toHaveBeenCalledTimes(dataMock.length);
+        expect(labelMock).toHaveBeenCalledWith(expected);
+      });
+    });
+
+    describe('provided as function returning an element', () => {
+      it('render returned elements', () => {
+        const { container } = render({
+          label: (props) => (
+            <text key={props.dataIndex}>{props.dataIndex}</text>
+          ),
+        });
+
+        container.querySelectorAll('text').forEach((label, index) => {
+          expect(label).toHaveTextContent(index);
+        });
+      });
+    });
+
+    describe('provided as element', () => {
+      it('renders with expected props', () => {
+        const ComponentMock = jest.fn(() => null);
+        render({
+          label: <ComponentMock />,
+        });
+
+        expect(ComponentMock).toHaveBeenCalledTimes(dataMock.length);
+        expect(ComponentMock).toHaveBeenCalledWith(
+          expectedNormalizedDataMock,
+          {}
+        );
+      });
+    });
+  });
+
+  describe('"labelStyle"', () => {
+    it('assign provided value to each label as className', () => {
+      const { container } = render({
+        label: true,
+        labelStyle: { pointerEvents: 'none' },
+      });
+
+      container.querySelectorAll('text').forEach((label) => {
+        expect(label).toHaveStyle('pointer-events: none');
+      });
+    });
+  });
+});
