+++ src/__tests__/testUtils/getArcInfo.js
@@ -1,23 +1,20 @@
 import parseSVG from 'svg-path-parser';
 import { degrees as getDegrees } from '@schwingbat/relative-angle';
 
-const CENTER = {
+export const DEFAULT_ORIGIN = {
   x: 50,
   y: 50,
 };
 
-export function getArcInfo(element) {
-  return getArcPathInfo(element.getAttribute('d'));
-}
-
-export function getArcPathInfo(d) {
+export function getArcInfo(element, { arcOrigin = DEFAULT_ORIGIN } = {}) {
+  const d = element.getAttribute('d');
   const [moveto, arc] = parseSVG(d);
 
   if (arc.rx !== arc.ry) {
     throw new Error('Provided path is not the section of a circumference');
   }
 
-  let degrees = getDegrees(CENTER, arc) - getDegrees(CENTER, moveto);
+  let degrees = getDegrees(arcOrigin, arc) - getDegrees(arcOrigin, moveto);
   const degreesAbsolute = Math.abs(degrees);
 
   if (
@@ -32,8 +29,9 @@ export function getArcPathInfo(d) {
       x: moveto.x,
       y: moveto.y,
     },
-    startAngle: getDegrees(CENTER, moveto),
+    startAngle: getDegrees(arcOrigin, moveto),
     lengthAngle: degrees,
     radius: arc.rx,
+    origin: arcOrigin,
   };
 }
