+++ src/__tests__/testUtils/index.js
@@ -1,0 +1,7 @@
+export {
+  render,
+  dataMock,
+  expectedNormalizedDataMock,
+  PieChart,
+} from './render';
+export { getArcInfo, DEFAULT_ORIGIN } from './getArcInfo';
